ci(node): run examples in CI
