+++ nodejs/examples/jest.config.cjs
@@ -1,0 +1,6 @@
+/** @type {import('ts-jest').JestConfigWithTsJest} */
+module.exports = {
+    preset: "ts-jest",
+    testEnvironment: "node",
+    testPathIgnorePatterns: ["./dist"],
+  };
